Add new feature in game method.

+++ src/main/Opponent.java
@@ -2,19 +2,19 @@ package main;
 
 public class Opponent
 {
-    private int pointsLife;
+    private double pointsLife;
 
     public Opponent()
     {
         this.pointsLife = 100;
     }
 
-    public int getPointsLife()
+    public double getPointsLife()
     {
         return pointsLife;
     }
 
-    public void setPointsLife(int pointsLife)
+    public void setPointsLife(double pointsLife)
     {
         this.pointsLife = pointsLife;
     }
